+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+#### 1.1.0:
+- Added a hotkey to toggle AI
+- Added the ability to directly add cards to players/bots in the card spawn menu
+
 #### 1.0.0:
 - Released